feat: SFT 7GPU 설정 추가, HF 모델 업로드 스크립트 및 tb2-lite 평가 결과 업데이트
- Qwen3.5-9B 7GPU 풀파인튜닝 설정 추가
- HuggingFace 모델 스테이징/업로드 스크립트 추가
- SFT 학습 스크립트 DDP 안정성 개선 및 문서 업데이트
- tb2-lite Qwen3.5 4B/9B 체크포인트 및 파이널 결과 추가
- unsloth 컴파일드 캐시 동기화

+++ unsloth_compiled_cache/BatchNorm3d.py
@@ -118,4 +118,4 @@ def forward(self, input: Tensor) -> Tensor:
         bn_training,
         exponential_average_factor,
         self.eps,
-    ).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype)
+    ).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype)
