+++ unsloth_compiled_cache/LayerNorm.py
@@ -73,4 +73,4 @@ from transformers.models.qwen3_5.modeling_qwen3_5 import (F)
 def forward(self, input: Tensor) -> Tensor:
     return F.layer_norm(
         input, self.normalized_shape, self.weight, self.bias, self.eps
-    ).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype)
+    ).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype)
